How to add a new resource and node type
An experiment to add the reasoning engine and prompt node. This process needs to be improved so that it is easy to add this stuff. I imagine this will be a common thing that people will want to do. Every person has their own idea of what abstractions should exist, they should be able to build those.