nodes and resources do not depend on each other
There was a risk of an import cycle because the distinction between what a node and what a resource were supposed to do was confusing. This isn't perfect, but I think this is closer to what this should look like. I think moving the types into `graph` feels right. I don't know what the name should be exactly, but it is something like this.